+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change Log
 ## next
 
 * `TraktV2`: add `isUnauthorized(response)`, `isAccountLocked(response)` and `isNotVip(response)` helper methods.
+* `TraktV2`: add `getPageCount(response)` and `getItemCount(response)` helper methods.
 
 ## 6.15.0
 _2024-08-30_